Environment script code for generating directories moved to a function
This is to make sure that this code is not run every time the script is run (since it only ever needs to be run once per project anyway). Also, made the houdini_env.sh script open houdinifx instead of houdini by default. Fixed a couple typos in README.md. Lastly, deleted a symlink that pointed to itself.