Removed selection of node from asset manager code
Now all the functions in the asset manager expect a node as an
input parameter. Also, started work on Nuke tools; created a Nuke
launcher script, code to create Nuke jobs on the farm and a Nuke
ROP otl to send those jobs to the farm. Started generic code to
send jobs to the farm for Maya as well.

Also, did some more abstracting of UI code in the code base.