Added new Hair Ubershader otl asset
The default shader that comes with the fur asset has a few glaring
issues. The biggest of which is the shading model is (surprisingly)
not physcially based! Thus the hair ubershader uses the physically
based hair model VOP. The diffuse and specular were renamed to
primary and secondary reflection. It also has promoted options for
all the missing parameters including transmission (although by
default it just uses the colors from primary reflection).