Added "Issue Submission" tool to open an online submission form
When starting apps, made cwd the USER_DIR (instead of $JOB/tmp).
Changed the name and title of newOTL to new (in preparation for
it to handle scenes as well as assets). Added ISSUE_URL envar.
Fixed getNodeInfo tool to not throw an exception when an asset
node is not selected.